Initial working prototype of new universe time clock.

+++ Engine/Runtime.Abstractions/Source/IStopwatch.cs
@@ -8,12 +8,12 @@
 
         void Reset();
 
-        int GetHours();
+        long GetHours();
 
-        int GetMinutes();
+        long GetMinutes();
 
-        int GetSeconds();
+        long GetSeconds();
 
-        int GetMilliseconds();
+        long GetMilliseconds();
     }
 }
